backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-dateutil==2.9.0.post0
feedparser==6.0.11
tenacity==9.0.0
qdrant-client==1.12.1
opensearch-py==2.8.0
google-generativeai==0.8.3
litellm==1.60.2
apscheduler==3.10.4
email-validator==2.2.0
boto3==1.35.70
redis==5.2.1
celery==5.4.0
python-dotenv==1.0.1
orjson==3.10.12
numpy==2.1.3
python-multipart==0.0.17
